Fix ec_setprop test on Chrome
As written, this test assumes an additional escape when retrieving the
escaped image URL from a CSS property vs when you retrieve it from an
image src attribute.  While this seems to be true in PhantomJS, it's not
happening in Chrome or Firefox.

Since the spirit of the test seems to be checking that appropriate
escapes are applied at all, I'm loosing the assertions to accept either
escaped version, so that the test will pass on both Chrome and Phantom.