SchoolAutocompleteDropdownTest passes on Chrome
This component uses Fetch to send search queries to the server.  Fetch
isn't stubbed by sinon's fake server.  This was passing in Phantom
because Phantom doesn't implement fetch, and falls back on a polyfill
that uses XMLHttpRequest, which _is_ stubbed by sinon's fake server.
But running in Chrome we use real Fetch.

The solution here is to stub fetch directly instead of a using a fake
server.  That's a pretty significant change but I think we ended up in a
good place.  Also cleaned up some tests here to use async/await which
allows them to read in order.