Reduce the footer content.
When working in code studio, there is a small, grey footer in the bottom left where students can select their language, view the privacy policy, review the terms of service, translate content, get support, or report abuse. This footer can be problematic though because on some small screen it will block critical buttons or the footer will be too small and can't display all of it's content. We also think that it makes sense for this content to be in the toolbar menu at the top of the screen anyways.