PL: Turn on legacy survey summaries view for local summer
As a followup to what https://github.com/code-dot-org/code-dot-org/pull/34654 did for CSF Intro, this shows legacy facilitator survey summaries for CSD/CSP local summer workshops.

It also adjusts the page so that we only show summaries when there are results for that user.  Additionally, the CSF Intro summary for all workshops is only shown if the user has a result for their own CSF Intro workshops.