Fixed encoding issue for student name in header
We cached a student's display name in a cookie so it is easily
accessible by front-end javascript. But if a student has
accented latin characters in their name, then their name will be URI
encoded in the cookie. When we were displaying the name stored in the
cookie, we were incorrectly treating it as an HTML escaped String. This
change uses URI decoding instead.