Check password confirmation when setting/updating password
On the "My Account" page, users can create a personal login or update
their account's existing password. There is already a "Password
confirmation" field where the user is supposed to type in their password
again. However, on the backend we were not verifying that the password
and password confirmation fields matched.

This change makes sure we run password field validations when a user is
trying to create or change their password.