Correct ages in our privacy policy
Updates our privacy policy to correctly reflect our system's behavior
when a student signs one of our petitions.  Before, our policy was
inconsistent: In some places it said we anonymize petition signatures
for students under age 16, and in other places it says we anonymize them
for students under age 13.

This commit updates all language to say we anonymize submissions under
age 16, which is what our system actually does, as of a policy change in
May 2018 to comply with GDPR.