Add afe/submit page
Return from sign-up/sign-in to /afe/submit instead of /afe.  This page
only loads the eligibility form, which in turn should only perform the
final submit and redirect at this point.  Provided a message indicating
that the request is being processed.

This isn't great - what I'd really like is for all relevant state to be
persisted in some way our server has access to and for our sign-up route
to redirect to a submit route, which in turn performs the submission and
then returns a redirect to a success or error page.  However, we've
built a system that keeps this temporary state on the client - which
might make more sense given how messy persisting state across sign-in
cna be - and I'm going to try and wrestle it into shape.