Deleting an application deletes its unsent email
Every once in a while, our application email sending process hits an
error where it tries to send an email for an application that has been
deleted. (example:
https://app.honeybadger.io/projects/45435/faults/48199539)  This isn't
terribly common because we don't often delete applications and it has to
be deleted within a certain window to hit this error as well.  However,
it would be best if our system cleaned up unsent emails automatically
when an application was deleted.  This change adds that behavior.

I've added a unit test to cover this behavior, and am leaning on
existing tests to catch any edge cases I haven't considered.