Best-effort titleize school name and address
Makes an effort to correct capitalization of the school name, school
district name, and school address fields before sending them to CSTA, at
CSTA's request.  We typically pull this information out of our NCES
records in upcase.  This is never going to be a perfect conversion, but
it's likely to be good enough for their needs.