Fix launching import on load
The import flow wasn't lauching when returning from reauthorization
because the hook to do this wasn't setting a roster provider.

This change fixes that, and also moves ownership for dispatching this
action to the "top level" TeacherHomepage component, so that we aren't
plumbing a queryparam through several layers and then expecting a
component down the tree to own an "onload" action that shows a modal
dialog.  I know this centralizes some responibility, but in this case
that seems like a good idea since it's an action that dominates the user
experience and we don't want multiple modals fighting with each other.