Fix ForbiddenAttributesError in registrations#new
Resolves [this Honeybadger
error](https://app.honeybadger.io/projects/3240/faults/40094122) which
has seen tens of incidents since June 1, 2020.

It seems this `ActiveModel::ForbiddenAttributesError` can occur when a
PartialRegistration is in progress, and then the user clicks on a
`sign_up` link that includes a `user[user_type]` queryparam.  This
caused us to enter the finish registration case and initialize a user
with an ActionController::Parameters object that has not been through a
`permit` operation.

I've written a unit test reproduced the error in question, and updated
it to define a desired behavior in the above scenario: The user_type
queryparam should "win" over a user_type stored in the session.

To prevent the error, I'm calling `permit(:user_type)` on the parameters
object before we initialize a user with it.