PartialRegistration gets a smaller session cookie footprint
For years we have been battling with a low volume of
ActionDispatch::Cookies::CookieOverflow errors.  This error is caused by
storing too much data in our session cookie, which has a strict limit of
4KB (when encrypted).

We have poor session hygiene in a few places, but one of the most
obvious ones is in our sign-up flow, which receives a large portion of
these errors (more than one thousand per month). We can probably blame
this on our "partial registration" system, which builds a User object
from our first interaction with a user (say, an OAuth callback or the
first page of our registration form) then serializes it into the
session, retrieving it again when the user submits our "finish sign up"
form.  If you browse [a relevant error on
HoneyBadger](https://app.honeybadger.io/projects/3240/faults/51743266)
you will find a nearly-complete user attributes hash buried in the
session.

I say "nearly" because we've worked around a _worse_ version of this
issue before: Some OAuth providers use tokens that are large enough that
they were consistently causing this issue.  So we already had a system
for keeping these tokens in the shared application cache (memcached)
instead of the session cookie.  See #17106 and #25002 for more history
on this approach.

However, there's a lot more we can do to shrink the footprint of
PartialRegistration within the session.  For example, we don't need to
store sign_in_count or deleted_at for a user that hasn't been created
yet.  And there are potentially large fields like school and full_address that we can keep
in memcached out of an abundance of caution.  In total I've removed more than half of the user attributes were were storing in the session.

I've opted to block certain attributes instead of creating an
include-list here, in the interest of not making this a "gotcha" for a
future developer trying to update the registration flow.  I could see an
argument either way though - interested in input.

Reminder: This isn't the whole picture.  Other known large value in our
session is the "callouts_seen" list, and other smaller values contribute
toward the 4KB total as well.
It's the combination of these things that causes a CookieOverflow
error; each one we address will reduce the overall volume of these
issues, but they probably won't go away entirely until we've dealt with
both of them.