Move cancel tracking to RegistrationsController
Moves the responsibility for reporting a sign-up cancellation through SignUpTracking out of PartialRegistration and into RegistrationsController.

This feels right - SignUpTracking always seemed a little out-of-scope for PartialRegistration anyway, we just put it here for easy access to the provider name.

I'm making this refactor in preparation for another change which will cancel/clear a PartialRegistration upon successful sign-up, which shouldn't result in tracking a cancel event.

I haven't added a new test for this because there was already excellent coverage of this flow (including the sign-up tracking) in GoogleOauth2Test.  I've modified that test with a couple of new assertions checking the expected PartialRegistration.in_progress? value.