Add a download link to the Aloe Blacc text compression video
This version of the Aloe Blacc text compression video on our videos page (https://code.org/educate/resources/videos) is shorter than the version we use in our curriculum, and did not have a fallback version or download link.  We've just uploaded this version of the video to videos.code.org and need to update the code here to make the appropriate download link appear.