Remove Hound
We haven't used Hound to lint our Ruby code in a long time - that integration hasn't been working in years.  This change removes the CI configuration for Hound along with a couple of documentation references to it that are out-of-date.  (Note: There are some other out-of-date details adjacent to this change in TESTING.md, but I'm not ready to tackle that yet.)

I haven't removed .rubocop_hound.yml because it is actually used, imported as part of .rubocop.yml.  We could carefully merge these, but it might be better for someone to take a fresh look at our ruby style enforcement anyway.