Extract international partners table into untranslated view
Turns out that Crowdin _super_ doesn't play nicely with markdown tables; this page is currently very broken in translated languages. As a short-term fix, we simply extract the table from the page into a non-translated view. As future follow-up work, we can look into translating certain parts of the view, like the headers or country names.