"Hide Toolbar" Helper updates
In summary: we no longer lose the helper when rotating an iPhone from landscape to portrait in an AI for Oceans level.

In an AI for Oceans level, when rotating an iPhone from landscape to portrait, the helper was being hidden, and would no longer be seen once rotated back to landscape, nor indeed for another year due to the cookie being set.  This was because we were getting both an "orientationchange" event and a "resize" event, and the heuristic for determining whether the toolbar was hidden thought it had been hidden in the first event (specifically, window.innerHeight was equal to document.body.offsetHeight), while it wouldn't have thought it was dismissed in the second event.

We only need to respond to the second, "resize", event, and things work better.  A regular Flappy level appears to continue working.

This also adds support for iOS 14.

Followup to https://github.com/code-dot-org/code-dot-org/pull/34390