Reload user after attempting takeover
For some reason, this is necessary now. It seems to be ultimately caused
by this change https://github.com/heartcombo/devise/pull/4796 which
stopped validating the user on save after updating the sign-in tracking
fields. With that change in place, after calling `lookup_user.update!`
earlier in this method the updated values - including the `provider` and
`uid` - stick around on the user object that's in memory. Manually
reloading the user object after a failed update resolves this.