Oceans: revisiting responsiveness
This revisits the code-based responsiveness that was begun last year.  As well as addressing a bunch of the code review feedback from before, it takes a slightly new approach.

Knowing that the "codeApp" div gives us the allocated playground, we simply attempt to fill it, with constraints.  However, we do one extra thing for Oceans: if the window width is smaller than codeApp, then we shrink ourselves down.  Because we are positioned absolutely, we won't have an impact on the space allocated.  But it does mean that we will keep the tutorial small enough to fit on screen.