Replace alias_method_chain in StiFactory
Since it has been deprecated. The new reccommended process is to use `prepend`, but we're doing some funky stuff here, and it's not clear to me that is actually what we want to do. Specifically, in the `new_with_factory` method we are explicitly calling `klass.new_without_factory`, and I'm not sure of the best way to reimplement that pattern when using `prepend`.

Instead, I'm simply unwrapping `alias_method_chain`, and using the underlying `alias_method` method instead.