Use generic - rather than region-specific - regional partner in test
It's not 100% clear why this is necessary, but here's what we do know:

- https://github.com/rails/rails/pull/33673 changed something about the way child records are persisted when created as a part of their parents.
- Regional Partner Mappings should be unique, and this Regional Parter was invalid as created, because a Regional Partner for Alabama already exists in our test database as a result of seeding.
- On Rails 5.2, attempting to join sessions to workshops with `regional_partner.pd_workshops.joins(:sessions` fails when `regional_parter` is invalid.

Like I said, it's not entirely clear why these things are connected but a `git bisect` indicates that they are. A simple - if uninformed - fix is to simply not use a regional partner mapping here when that's not what we are actually testing.