Remove 'escaped cookies' test
The latest version of Rack includes a change to this behavior (https://github.com/rack/rack/commit/5ccca4722668083732ea2d35c56565fcc25312f8) which caused this test to start failing. We could simply update the test to conform to the new behavior, but because we don't actually rely on this behavior anywhere and to reduce future potential noise, it seems better to just remove the test.