Partially defend against Tabnapping in Button component
The Button component supports the use of `target="_blank"` to open links in a new tab, which is an inherently insecure operation. I looked into removing this functionality, but we unfortunately use it in a couple of places. Enough that it would take more work to remove than I have time for right now, so I instead chose to add in what defenses are possible in this scenario.