Add a unique index to ProgrammingEnvironment#name
Mostly so that we can uniquely identify ProgrammingExpressions by `ProgrammingExpression#key` and `ProgrammingEnvironment#name`, but also because it's generally nice to have at least one non-id field be unique.

Currently we only have four programming environments whose names do not collide, so I'm not worried about this change being expensive to implement or introducing the danger of invalidating existing environments.