Add 'Code Block' button to more markdown editors
Specifically, to the lesson Overview, Preparation, and Purpose field editors.

Also, the editor for the actitivity secion markdown field is starting to get a _lot_ of buttons; we're going to want to make some changes there (and probably also do something to unify the editor we use for that with the one we use for these fields), so to start playing around with some ideas for what we could do I've collapsed the two-now-three buttons that we had in this editor down to a single dropdown! I think this will work much better, but either way doing this will let us get some early feedback from our editors.