Restrict Resource PDF generation to just PDFs and Google Docs
Specifically, remove support for Google Slides, Google Sheets, etc.

Existing CurriculumBuilder functionality has similar restrictions, and support for Google Slides in particular was causing several issues. We could consider reenabling support for them in the future, but should be aware that it will likely take some special-casing to do so.