Speed up Lesson#unplugged_lesson?
Currently, this method is implemented by querying for all ScriptLevels
in the Lesson's Script, and manually filtering them down to just the one
ScriptLevel we actually care about. We should instead query directly for
the ScriptLevel we care about.

Right now, calling Script#summarize invokes this method quite a lot.
Specifically:

- Script#summarize_for_unit_edit invokes LessonGroup#summarize_for_unit_edit for each LessonGroup in the Script
- LessonGroup#summarize_for_unit_edit invokes Lesson#summarize_for_unit_edit for each Lesson in the LessonGroup
- Lesson#summarize_for_unit_edit invokes Lesson#summarize
- Lesson#summarize invokes ScriptLevel#summarize for each ScriptLevel in the Lesson
- ScriptLevel#summarize invokes ScriptLevel#level_display_text
- ScriptLevel#level_display_text invokes Lesson#unplugged_lesson? for the Lesson associated with the ScriptLevel

What this means is that when we call Script#summarize, we are querying
for _all_ ScriptLevels in the entire Script not just once, but _once for
every ScriptLevel in the entire Script_.