Localize Resource URLs
Usually we want to avoid allowing translators to edit url values, but for Resources this is actually the system we plan to use to allow international partners to create translated versions of resource content. The international team plans to own the work to make sure these fields are secured in Crowdin against malicious translator activity.