Remove unnecessary manual 'create' invocation
We don't want to invoke the create operation every time we seed, because most of those seeds are going to be updates, not creates. Unfortunately, activerecord-import doesn't give us a great way to differentiate creates from updates; fortunately, we don't actually have anything dependent on the create hook right now. The simple solution is to just not invoke the hook.