Fixtures to Factories: Script Levels Controller
Also Hint View Requests Controller.

I think I finally figured out a decent way to accomodate our need to continue to support fixture-provided Scripts without also providing levels in the fixture. Specifically, we have a number of tests which are examining functionality specific to certain scripts, and so need to use those specific scripts and not a factory-created one in the test itself.  Once we remove the levels fixture, those scripts will still be seeded, but will be seeded without levels.

The simple fix for that subset of tests that care both about scripts and about levels is to simply use our factory to create levels for those scripts when they're missing. It's not nearly as clean as it would be to remove the scripts fixture as well and do everything in a factory, but I think it will work just fine as a stepping stone.