Add retries to Resource PDF Generation
For background, we've had an intermittent issue with the Curriculum PDF generation process in which when it comes time to collate all our individual PDFs together in the rollup, GhostScript complains that it can't find all the files that we're telling it to roll up, and the PDF generation fails.

I've been able to track down the failure to specifically the title pages that we generate in the Resource rollup to differentiate the lessons within a script. Unfortunately, I have been unable to get any kind of reliable repro, much less to figure out *why* this is happening in the first place.

Because I know where the problem is but not why it is, and because it happens so rarely as to make further investigation difficult, I'm resorting to a blind retry. It's not a good solution, but it's the best I got at this point.