Template partials get parameters
Template types that use the `{{ }}`` syntax for rendering a partial can now pass parameters to those partials.  The syntax is like this:

```
{{ path/to/partial, param1: "value1", param2: "value2" }}
```

The motivation for this work is that somebody can author a single file (say an `.md.partial`) which includes partials that take user-facing strings as parameters, and the file can be localised.