Pegasus: Support for HTML-only partials
This is an idea to allow partials to take parameter strings which will be rendered for the user, but that avoid any risk of a malicious translation calling into our code server-side.

Now, a `{{ }}` partial that takes parameters must have `.html` source code, and at render time, we replace each `%key_value%` in the HTML with the desired parameter.