Improve code.org/ layout & remove SVG support for now
Unfortunately, we don't get the helpful layout support of background-image when we have an SVG that we want to draw ourselves, as far as I can tell.  We want the left & right images in the hero to just size themselves into available space, rather than increasing the height of the flexbox, but I wasn't able to figure out how to do that yet.